replace QCache by QMap
QCache cannot be used to keep track of images stored
in the nvg context.It will evict data at any time
without notie and the image in the store is lost
but uses memory. For image name to id mapping i replace
QCache  it by a QMap.
A better cache solution has to be found.