share QNanoPainter between instance of QNanoItemPainter
as QNanoPainter owns NVG Context,it is also shared,
incuding stored fonts and images.
I simply made a private get function which returns a
QSharedPointer<QNanoPainter> to the same instance, every
time it is called.
Instead of new, this function is called now  and the member m_painter
is also a QSharedPointer<QNanoPainter>.
I extended the QNanoPainter::reset to reset its internal members too
, when it is called.
I need to reset the state of the painter (and the context) in
QNanoQuickItemPainter::render(),so it has a defined state,
when the next QNanoQuickItemPainter uses it for rendering.